# backend/bracket/models/db/user.py
from __future__ import annotations

from typing import TYPE_CHECKING, Annotated

from heliclockter import datetime_utc
from pydantic import BaseModel, StringConstraints, field_validator

from bracket.models.db.account import UserAccountType
from bracket.models.db.shared import BaseModelORM
from bracket.utils.id_types import UserId

if TYPE_CHECKING:
    from bracket.logic.subscriptions import Subscription

# Un usuario inactivo no puede autenticarse (ni /token ni JWT ya emitidos).
class UserBase(BaseModelORM):
    email: str
    name: str
    created: datetime_utc
    account_type: UserAccountType
    active: bool = True

    @property
    def subscription(self) -> Subscription:
        from bracket.logic.subscriptions import subscription_lookup

        return subscription_lookup[self.account_type]


class UserInsertable(UserBase):
    password_hash: str | None = None


class User(UserBase):
    id: UserId
    password_hash: str | None = None


class UserPublic(UserBase):
    id: UserId


class UserToUpdate(BaseModel):
    email: str
    name: str


class UserPasswordToUpdate(BaseModel):
    password: Annotated[str, StringConstraints(min_length=8, max_length=48)]


class DemoUserToRegister(BaseModelORM):
    captcha_token: str


class UserToRegister(BaseModelORM):
    email: str
    name: str
    password: str
    captcha_token: str


class UserInDB(UserBase):
    id: UserId
    password_hash: str


class UserToCreateByAdmin(BaseModelORM):
    """Alta de usuario desde administracion (no expuesta publicamente)."""

    email: str
    name: str
    password: Annotated[str, StringConstraints(min_length=8, max_length=48)]
    account_type: UserAccountType = UserAccountType.REGULAR

    @field_validator("account_type")
    @classmethod
    def _only_regular_or_admin(cls, value: UserAccountType) -> UserAccountType:
        if value not in (UserAccountType.REGULAR, UserAccountType.ADMIN):
            raise ValueError("account_type should be REGULAR or ADMIN")
        return value


class UserActiveToUpdate(BaseModelORM):
    active: bool


class UserAccountTypeToUpdate(BaseModelORM):
    account_type: UserAccountType

    @field_validator("account_type")
    @classmethod
    def _only_regular_or_admin(cls, value: UserAccountType) -> UserAccountType:
        if value not in (UserAccountType.REGULAR, UserAccountType.ADMIN):
            raise ValueError("account_type should be REGULAR or ADMIN")
        return value